plotting/requirements.txt
pandas==2.3.3
matplotlib==3.10.8
numpy==2.4.0
PyYAML==6.0.3

